fix: MapView not displayed
