+++ src/App.tsx
@@ -1,7 +1,10 @@
+import { useEffect } from "react";
 import { Menu } from "./components/menu";
 import { handleDragOver, handleDrop } from "./loadCSV";
+import { setupMapView } from "./mapView";
 
 export const App = () => {
+  useEffect(setupMapView, []);
   return (
     <>
       <Menu></Menu>
